+++ git-extensions/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 0.1.0
 
+ - added native github client (python)
  - finish now checks if current branch contains master
  - add research command to git devbliss
  - create, finish and delete branches and hotfixes